# payload/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Egregore Projection Plane",
    "version": "0.6.0-phase1"
  },
  "paths": {
...
    "/api/v1/ombudsman/cells/{cell_id}": {
      "get": {
        "tags": [
          "ombudsman"
        ],
        "summary": "Get Cell",
        "description": "Return detailed state for a single cell.",
        "operationId": "get_cell_api_v1_ombudsman_cells__cell_id__get",
        "parameters": [
          {
            "name": "cell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cell Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Cell Api V1 Ombudsman Cells  Cell Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ombudsman/route/{taxonomy}": {
      "get": {
        "tags": [
          "ombudsman"
        ],
        "summary": "Route To Cell",
        "description": "Find the best available cell for a given taxonomy path.",
        "operationId": "route_to_cell_api_v1_ombudsman_route__taxonomy__get",
        "parameters": [
          {
            "name": "taxonomy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taxonomy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Route To Cell Api V1 Ombudsman Route  Taxonomy  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ombudsman/cells/{cell_id}/load": {
      "post": {
        "tags": [
          "ombudsman"
        ],
        "summary": "Update Load",
        "description": "Update a cell's load index (called by the cell or monitor).",
        "operationId": "update_load_api_v1_ombudsman_cells__cell_id__load_post",
        "parameters": [
          {
            "name": "cell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cell Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CellLoadUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Load Api V1 Ombudsman Cells  Cell Id  Load Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ombudsman/dispatch": {
      "post": {
        "tags": [
          "ombudsman"
        ],
        "summary": "Dispatch",
        "description": "Route a request to matching cells, execute them, and fuse the streams.",
        "operationId": "dispatch_api_v1_ombudsman_dispatch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Dispatch Api V1 Ombudsman Dispatch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ombudsman/advisory/call": {
      "post": {
        "tags": [
          "ombudsman"
        ],
        "summary": "Advisory Call",
        "description": "Let one cell invoke another cell and return the resulting RFE stream.",
        "operationId": "advisory_call_api_v1_ombudsman_advisory_call_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvisoryCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Advisory Call Api V1 Ombudsman Advisory Call Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ombudsman/dispatch_from_orchestrator": {
      "post": {
        "tags": [
          "ombudsman"
        ],
        "summary": "Dispatch From Orchestrator",
        "description": "Accept an orchestration-suite JobRequest and dispatch it as a WorkUnit.",
        "operationId": "dispatch_from_orchestrator_api_v1_ombudsman_dispatch_from_orchestrator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Job Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Dispatch From Orchestrator Api V1 Ombudsman Dispatch From Orchestrator Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvisoryCallRequest": {
        "properties": {
          "caller_cell_id": {
            "type": "string",
            "title": "Caller Cell Id"
          },
          "target_cell_id": {
            "type": "string",
            "title": "Target Cell Id"
          },
          "input": {
            "type": "string",
            "minLength": 1,
            "title": "Input"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "caller_cell_id",
          "target_cell_id",
          "input"
        ],
        "title": "AdvisoryCallRequest"
      },
...
      "CellLoadUpdate": {
        "properties": {
          "cell_id": {
            "type": "string",
            "title": "Cell Id"
          },
          "load_index": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Load Index"
          }
        },
        "type": "object",
        "required": [
          "cell_id",
          "load_index"
        ],
        "title": "CellLoadUpdate"
      },
      "DispatchRequest": {
        "properties": {
          "taxonomy": {
            "type": "string",
            "title": "Taxonomy",
            "description": "Target taxonomy path."
          },
          "input": {
            "type": "string",
            "minLength": 1,
            "title": "Input",
            "description": "Primary natural-language request."
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "advisory_call": {
            "type": "boolean",
            "title": "Advisory Call",
            "description": "Enable advisory calls to linked cells.",
            "default": false
          },
          "fuse": {
            "type": "boolean",
            "title": "Fuse",
            "description": "Run RFE fusion over emitted streams.",
            "default": true
          },
          "constraints": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints",
            "description": "Optional RFE constraints."
          }
        },
        "type": "object",
        "required": [
          "taxonomy",
          "input"
        ],
        "title": "DispatchRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}